refactor: fixed typo in package name
