--- petersbugredu_wrap/__init__.py
@@ -1,5 +1,0 @@
-"""
-Library for accessing petersburg education website via python
-"""
-from petersbugredu_wrap.client import Client
-import petersbugredu_wrap.types
