--- petersbugredu_wrap/types/__init__.py
@@ -1,7 +1,0 @@
-"""
-Module used for storing types of this library
-"""
-from petersbugredu_wrap.types.child import Child
-from petersbugredu_wrap.types.teacher import Teacher
-from petersbugredu_wrap.types.action_payload import ActionPayload
-from petersbugredu_wrap.types.identity import Identity
