+++ petersburgedu_wrap/__init__.py
@@ -1,0 +1,5 @@
+"""
+Library for accessing petersburg education website via python
+"""
+from petersburgedu_wrap.client import Client
+import petersburgedu_wrap.types
