+++ petersburgedu_wrap/client.py
@@ -1,11 +1,10 @@
-import json
 import logging
 import requests
 
-from petersbugredu_wrap.types.education import Education
-from petersbugredu_wrap.utils import endpoints, request_parameters
-from petersbugredu_wrap.errors.invalid_login_or_password_exc import InvalidLoginOrPasswordException
-from petersbugredu_wrap.types import Child, ActionPayload, Identity
+from petersburgedu_wrap.types.education import Education
+from petersburgedu_wrap.utils import endpoints, request_parameters
+from petersburgedu_wrap.errors.invalid_login_or_password_exc import InvalidLoginOrPasswordException
+from petersburgedu_wrap.types import Child, ActionPayload, Identity
 
 
 class Client:
