+++ petersburgedu_wrap/types/__init__.py
@@ -1,0 +1,7 @@
+"""
+Module used for storing types of this library
+"""
+from petersburgedu_wrap.types.child import Child
+from petersburgedu_wrap.types.teacher import Teacher
+from petersburgedu_wrap.types.action_payload import ActionPayload
+from petersburgedu_wrap.types.identity import Identity
